# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API exposing CV content (about, basics, work, etc.) as JSON.",
        "title": "CV API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/achievements": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "summary": "All work achievements grouped by company",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.AchievementsEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/work": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "summary": "Work experience",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WorkEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/work/achievements": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "summary": "All work achievements grouped by company",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.AchievementsEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.AchievementsEntity": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.CompanyAchievements"
                    }
                }
            }
        },
...
        "api.CompanyAchievements": {
            "type": "object",
            "properties": {
                "highlights": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.WorkEntity": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resume.Work"
                    }
                }
            }
        },
...
        "resume.Work": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "highlights": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}